requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
google-cloud-storage>=2.14.0
redis>=5.0.0
python-dotenv>=1.0.0
